refactor(frontend): consolidate components, services, and lib utilities
