--- frontend/src/app/evolution/_components/index.ts
@@ -1,2 +1,0 @@
-export { ChapterSummary } from "./ChapterSummary";
-export { TimelineItem } from "./TimelineItem";
