+++ frontend/src/components/form/CoverForm.tsx
@@ -1,7 +1,13 @@
 "use client";
 
 import { useForm } from "@/hooks";
-import { lato, montserrat, openSans, playfairDisplay, roboto } from "@/lib";
+import {
+	lato,
+	montserrat,
+	openSans,
+	playfairDisplay,
+	roboto,
+} from "@/lib/utils";
 import { CoverFormControls } from "./CoverFormControls";
 import { CoverPreviewDisplay } from "./CoverPreviewDisplay";
