--- frontend/src/lib/index.ts
@@ -1,15 +1,0 @@
-export {
-	downloadImage,
-	type FileSystemFileHandle,
-	type FileSystemWritableFileStream,
-	type SaveFilePickerOptions,
-} from "./download";
-export {
-	fontFamilyMap,
-	lato,
-	montserrat,
-	openSans,
-	playfairDisplay,
-	roboto,
-} from "./fonts";
-export { calculatePreviewDimensions, cn } from "./utils";
