--- frontend/src/services/analytics.ts
@@ -1,10 +1,0 @@
-/**
- * Client-side function to fetch analytics data
- */
-export async function getAnalytics() {
-	const response = await fetch("/api/analytics");
-	if (!response.ok) {
-		throw new Error(`Failed to fetch analytics: ${response.statusText}`);
-	}
-	return response.json();
-}
