--- frontend/src/services/cover.ts
@@ -1,111 +1,0 @@
-/**
- * Utility for generating cover images via the backend API
- */
-import type { ImageParams, ValidationError } from "@cover-craft/shared";
-
-export type { ImageParams };
-
-export type ApiErrorResponse = {
-	error: string;
-	details?: ValidationError[];
-};
-
-/**
- * Client-side function to call the generateImage endpoint and get blob
- */
-export async function generateImage(
-	params: ImageParams,
-): Promise<{ blob: Blob; clientDuration: number; duration?: number }> {
-	const startTime = performance.now();
-	const response = await fetch("/api/generateImage", {
-		method: "POST",
-		headers: {
-			"Content-Type": "application/json",
-		},
-		body: JSON.stringify(params),
-	});
-
-	const clientDuration = Math.round(performance.now() - startTime);
-
-	if (!response.ok) {
-		let errorBody: ApiErrorResponse | null = null;
-		try {
-			errorBody = await response.json();
-		} catch (_err) {
-			// ignore parse errors
-		}
-		const baseMessage = errorBody?.error ?? "Failed to generate cover image";
-		const details =
-			errorBody && Array.isArray(errorBody.details) ? errorBody.details : [];
-		let message = baseMessage;
-		if (details.length > 0) {
-			const detailsText = details
-				.map((detail) => `${detail.field}: ${detail.message}`)
-				.join("; ");
-			message = `${message}: ${detailsText}`;
-		}
-		const err = new Error(message) as Error & {
-			clientDuration?: number;
-			details?: ValidationError[];
-		};
-		err.clientDuration = clientDuration;
-		err.details = details;
-		throw err;
-	}
-
-	const blob = await response.blob();
-	return { blob, clientDuration };
-}
-
-export type JobStatusResponse = {
-	id: string;
-	status: "pending" | "processing" | "completed" | "failed";
-	progress: number;
-	total: number;
-	results: string[];
-	error?: string;
-	createdAt: string;
-	updatedAt: string;
-};
-
-/**
- * Submit a batch generation job
- */
-export async function generateBatchImages(
-	params: ImageParams[],
-): Promise<{ jobId: string }> {
-	const response = await fetch("/api/generateImages", {
-		method: "POST",
-		headers: {
-			"Content-Type": "application/json",
-		},
-		body: JSON.stringify(params),
-	});
-
-	if (!response.ok) {
-		let errorBody: ApiErrorResponse | null = null;
-		try {
-			errorBody = await response.json();
-		} catch (_err) {
-			/* ignore */
-		}
-		throw new Error(errorBody?.error ?? "Failed to submit batch job");
-	}
-
-	return await response.json();
-}
-
-/**
- * Poll for batch job status
- */
-export async function getBatchJobStatus(
-	jobId: string,
-): Promise<JobStatusResponse> {
-	const response = await fetch(`/api/jobStatus?jobId=${jobId}`);
-
-	if (!response.ok) {
-		throw new Error("Failed to fetch job status");
-	}
-
-	return await response.json();
-}
