--- frontend/src/services/index.ts
@@ -1,18 +1,0 @@
-export { getAnalytics } from "./analytics";
-export {
-	generateBatchImages,
-	generateImage,
-	getBatchJobStatus,
-	type ImageParams,
-	type JobStatusResponse,
-} from "./cover";
-export { type HealthResponse, health } from "./health";
-export {
-	type EventType,
-	type MetricPayload,
-	type MetricStatus,
-	sendDownloadEvent,
-	sendGenerateEvent,
-	sendMetrics,
-	type WcagLevel,
-} from "./metrics";
