--- frontend/src/services/metrics.ts
@@ -1,98 +1,0 @@
-import {
-	DOWNLOAD_CLICK_EVENT,
-	type EventType,
-	GENERATE_CLICK_EVENT,
-	type MetricPayload,
-	type MetricStatus,
-	type WcagLevel,
-} from "@cover-craft/shared";
-
-export {
-	DOWNLOAD_CLICK_EVENT,
-	type EventType,
-	GENERATE_CLICK_EVENT,
-	type MetricPayload,
-	type MetricStatus,
-	type WcagLevel,
-};
-
-/**
- * Client-side function to send metrics to the backend
- */
-export async function sendMetrics(data: Record<string, unknown>) {
-	const response = await fetch("/api/metrics", {
-		method: "POST",
-		headers: {
-			"Content-Type": "application/json",
-		},
-		body: JSON.stringify(data),
-	});
-
-	if (!response.ok) {
-		throw new Error(`Failed to send metrics: ${response.statusText}`);
-	}
-
-	return response.json();
-}
-
-/**
- * Client-side function to send generate cover event
- */
-export async function sendGenerateEvent(payload: Partial<MetricPayload>) {
-	try {
-		const payloadToSend: MetricPayload = {
-			event: GENERATE_CLICK_EVENT,
-			timestamp: new Date().toISOString(),
-			status: "success",
-			...payload,
-		};
-
-		if (
-			!payloadToSend ||
-			typeof payloadToSend !== "object" ||
-			Array.isArray(payloadToSend) ||
-			typeof payloadToSend.event !== "string"
-		) {
-			return;
-		}
-
-		await fetch("/api/metrics", {
-			method: "POST",
-			headers: { "Content-Type": "application/json" },
-			body: JSON.stringify(payloadToSend),
-		});
-	} catch (_err) {
-		// Silently fail - metrics are not critical to app functionality
-	}
-}
-
-/**
- * Client-side function to send download event
- */
-export async function sendDownloadEvent(payload?: Partial<MetricPayload>) {
-	try {
-		const payloadToSend: MetricPayload = {
-			event: DOWNLOAD_CLICK_EVENT,
-			timestamp: new Date().toISOString(),
-			status: "success",
-			...payload,
-		};
-
-		if (
-			!payloadToSend ||
-			typeof payloadToSend !== "object" ||
-			Array.isArray(payloadToSend) ||
-			typeof payloadToSend.event !== "string"
-		) {
-			return;
-		}
-
-		await fetch("/api/metrics", {
-			method: "POST",
-			headers: { "Content-Type": "application/json" },
-			body: JSON.stringify(payloadToSend),
-		});
-	} catch (_err) {
-		// Silently fail - metrics are not critical to app functionality
-	}
-}
